Rename namedfields to namedfieldsinit
